bundle workplace_location submodels together and move compute_logsums helper to util for sharing
